Test: Use native browser coverage and reduce Vitest exceptions
